FEAT local storage management for persistent data, REFACTOR styled naming convention
