+++ challenge-2/src/App.jsx
@@ -1,22 +1,25 @@
-import React, { useReducer } from 'react';
+import React, { useReducer, useEffect } from 'react';
 import { BrowserRouter as Router, Route, Routes } from 'react-router-dom';
-import { StyledBody, StyledPageWrapper } from './styled/Body';
+import { StyledBody, StyledPageWrapper } from './styled/Body.styled';
 import { NavBar, Notes, Archive } from './components';
 
 function reducer(notes, action) {
   switch (action.type) {
+    case 'set': {
+      const setNotes = action.payload ? action.payload : [];
+      return setNotes;
+    }
     case 'add':
-      console.log([...notes, action.payload]);
+      localStorage.setItem('notes', JSON.stringify([...notes, action.payload]));
       return [...notes, action.payload];
     case 'archive': {
-      console.log('archived called from dispatch');
       const updatedNotes = notes.map((note) => {
         if (note.id === action.payload.id) {
-          console.log(note);
           return { ...note, isArchived: !note.isArchived };
         }
         return note;
       });
+      localStorage.setItem('notes', JSON.stringify(updatedNotes));
       return updatedNotes;
     }
     default:
@@ -28,6 +31,11 @@ function App() {
   const [notes, dispatch] = useReducer(reducer, []);
   const activeNotes = notes.filter((note) => !note.isArchived);
   const archivedNotes = notes.filter((note) => note.isArchived);
+
+  useEffect(() => {
+    const localStorageNotes = localStorage.getItem('notes');
+    dispatch({ type: 'set', payload: JSON.parse(localStorageNotes) });
+  }, []);
 
   return (
     <Router>
