+++ challenge-2/src/styled/Body.styled.jsx
