+++ challenge-2/src/styled/NoteCard.styled.jsx
