+++ challenge-2/src/styled/NoteInput.styled.js
